+++ dots/.config/quickshell/ii/modules/ii/sidebarRight/quickToggles/ClassicQuickPanel.qml
@@ -32,6 +32,7 @@ AbstractQuickPanel {
         }
         NightLight {}
         GameMode {}
+        InstantReplay {}
         IdleInhibitor {}
         EasyEffectsToggle {}
         CloudflareWarp {}