+++ dots/.config/quickshell/ii/modules/ii/sidebarRight/quickToggles/androidStyle/AndroidToggleDelegateChooser.qml
@@ -43,6 +43,22 @@ DelegateChooser {
         onOpenMenu: root.openNightLightDialog()
     } }
 
+    DelegateChoice { roleValue: "instantReplay"; AndroidInstantReplayToggle {
+        required property int index
+        required property var modelData
+        buttonIndex: root.startingIndex + index
+        buttonData: modelData
+        editMode: root.editMode
+        gridRef: root.gridRef
+        expandedSize: modelData.size > 1
+        baseCellWidth: root.baseCellWidth
+        baseCellHeight: root.baseCellHeight
+        cellSpacing: root.spacing
+        cellSize: modelData.size
+        dropIndicatorRef: root.dropIndicatorRef
+        isUnused: root.isUnused
+    } }
+
     DelegateChoice { roleValue: "audio"; AndroidAudioToggle {
         required property int index
         required property var modelData